adding a warning incase someone calls provideOutputData(data) from an AndamaSource since this might lead to unexpected behavior (causing the node to stop processing further data).
The reason is that Sources don't have any input sources. Thus, provideOutputData will cause the status set to be completed AND getInputData() on the node will return null. As a result, then loop in the run method will end.
In contrast to all other nodes, Sources must manage their termination criterion themselves. This is caused by the nature of Sources not relying on other threads to produce any output.